+++ src/common/acl_util.c
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2019 Intel Corporation.
+ * (C) Copyright 2019-2020 Intel Corporation.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -50,6 +50,13 @@
  */
 #define PERM_READ_CH		'r'
 #define PERM_WRITE_CH		'w'
+#define PERM_CREATE_CONT_CH	'c'
+#define PERM_DEL_CONT_CH	'd'
+#define PERM_GET_PROP_CH	't'
+#define PERM_SET_PROP_CH	'T'
+#define PERM_GET_ACL_CH		'a'
+#define PERM_SET_ACL_CH		'A'
+#define PERM_SET_OWNER_CH	'o'
 
 /*
  * States used to parse a formatted ACE string
@@ -134,6 +141,27 @@ process_perms(const char *str, uint64_t *perms)
 			break;
 		case PERM_WRITE_CH:
 			*perms |= DAOS_ACL_PERM_WRITE;
+			break;
+		case PERM_CREATE_CONT_CH:
+			*perms |= DAOS_ACL_PERM_CREATE_CONT;
+			break;
+		case PERM_DEL_CONT_CH:
+			*perms |= DAOS_ACL_PERM_DEL_CONT;
+			break;
+		case PERM_GET_PROP_CH:
+			*perms |= DAOS_ACL_PERM_GET_PROP;
+			break;
+		case PERM_SET_PROP_CH:
+			*perms |= DAOS_ACL_PERM_SET_PROP;
+			break;
+		case PERM_GET_ACL_CH:
+			*perms |= DAOS_ACL_PERM_GET_ACL;
+			break;
+		case PERM_SET_ACL_CH:
+			*perms |= DAOS_ACL_PERM_SET_ACL;
+			break;
+		case PERM_SET_OWNER_CH:
+			*perms |= DAOS_ACL_PERM_SET_OWNER;
 			break;
 		default:
 			D_INFO("Invalid permission '%c'\n", str[i]);
@@ -423,6 +451,20 @@ daos_ace_to_str(struct daos_ace *ace, char *buf, size_t buf_len)
 		rc = write_char(&pen, PERM_READ_CH, &remaining_len);
 	if (perms & DAOS_ACL_PERM_WRITE)
 		rc = write_char(&pen, PERM_WRITE_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_CREATE_CONT)
+		rc = write_char(&pen, PERM_CREATE_CONT_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_DEL_CONT)
+		rc = write_char(&pen, PERM_DEL_CONT_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_GET_PROP)
+		rc = write_char(&pen, PERM_GET_PROP_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_SET_PROP)
+		rc = write_char(&pen, PERM_SET_PROP_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_GET_ACL)
+		rc = write_char(&pen, PERM_GET_ACL_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_SET_ACL)
+		rc = write_char(&pen, PERM_SET_ACL_CH, &remaining_len);
+	if (perms & DAOS_ACL_PERM_SET_OWNER)
+		rc = write_char(&pen, PERM_SET_OWNER_CH, &remaining_len);
 
 	return rc;
 }
