+++ src/common/tests/acl_util_tests.c
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 2019 Intel Corporation.
+ * (C) Copyright 2019-2020 Intel Corporation.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -282,23 +282,43 @@ test_ace_from_str_invalid_flags(void **state)
 }
 
 static void
-test_ace_from_str_perm_read_only(void **state)
+expect_perms_for_str(char *perms_str, uint64_t exp_perms)
 {
-	check_ace_from_valid_str("A::someuser@:r",
+	const char	*identity = "someuser@";
+	char		ace_str[DAOS_ACL_MAX_ACE_STR_LEN];
+
+	snprintf(ace_str, sizeof(ace_str), "A::%s:%s", identity, perms_str);
+
+	check_ace_from_valid_str(ace_str,
 				 DAOS_ACL_ACCESS_ALLOW,
 				 DAOS_ACL_USER,
 				 0,
-				 DAOS_ACL_PERM_READ,
-				 0, 0, "someuser@");
+				 exp_perms,
+				 0, 0, identity);
 }
 
 static void
-test_ace_from_str_perm_none(void **state)
+test_ace_from_str_perms(void **state)
 {
-	check_ace_from_valid_str("A::someuser@:",
-				 DAOS_ACL_ACCESS_ALLOW,
-				 DAOS_ACL_USER,
-				 0, 0, 0, 0, "someuser@");
+	expect_perms_for_str("", 0);
+	expect_perms_for_str("r", DAOS_ACL_PERM_READ);
+	expect_perms_for_str("w", DAOS_ACL_PERM_WRITE);
+	expect_perms_for_str("c", DAOS_ACL_PERM_CREATE_CONT);
+	expect_perms_for_str("d", DAOS_ACL_PERM_DEL_CONT);
+	expect_perms_for_str("t", DAOS_ACL_PERM_GET_PROP);
+	expect_perms_for_str("T", DAOS_ACL_PERM_SET_PROP);
+	expect_perms_for_str("a", DAOS_ACL_PERM_GET_ACL);
+	expect_perms_for_str("A", DAOS_ACL_PERM_SET_ACL);
+	expect_perms_for_str("o", DAOS_ACL_PERM_SET_OWNER);
+	expect_perms_for_str("rwcdtTaAo", DAOS_ACL_PERM_READ |
+					  DAOS_ACL_PERM_WRITE |
+					  DAOS_ACL_PERM_CREATE_CONT |
+					  DAOS_ACL_PERM_DEL_CONT |
+					  DAOS_ACL_PERM_GET_PROP |
+					  DAOS_ACL_PERM_SET_PROP |
+					  DAOS_ACL_PERM_GET_ACL |
+					  DAOS_ACL_PERM_SET_ACL |
+					  DAOS_ACL_PERM_SET_OWNER);
 }
 
 static void
@@ -550,8 +570,16 @@ test_ace_to_str_all_perms(void **state)
 			       DAOS_ACL_FLAG_ACCESS_FAIL,
 			       0,
 			       0,
-			       DAOS_ACL_PERM_READ | DAOS_ACL_PERM_WRITE,
-			       "L:F:EVERYONE@:rw");
+			       DAOS_ACL_PERM_READ |
+			       DAOS_ACL_PERM_WRITE |
+			       DAOS_ACL_PERM_CREATE_CONT |
+			       DAOS_ACL_PERM_DEL_CONT |
+			       DAOS_ACL_PERM_GET_PROP |
+			       DAOS_ACL_PERM_SET_PROP |
+			       DAOS_ACL_PERM_GET_ACL |
+			       DAOS_ACL_PERM_SET_ACL |
+			       DAOS_ACL_PERM_SET_OWNER,
+			       "L:F:EVERYONE@:rwcdtTaAo");
 }
 
 static void
@@ -658,7 +686,7 @@ check_ace_turns_back_to_same_str(const char *ace_str)
 static void
 test_ace_from_str_and_back_again(void **state)
 {
-	check_ace_turns_back_to_same_str("U:S:OWNER@:w");
+	check_ace_turns_back_to_same_str("U:S:OWNER@:rwcdtTaAo");
 	check_ace_turns_back_to_same_str("A:G:GROUP@:rw");
 	check_ace_turns_back_to_same_str("AUL:GS:somegroup@somedomain:rw");
 	check_ace_turns_back_to_same_str("AL:F:user1@:r");
@@ -812,8 +840,7 @@ main(void)
 		cmocka_unit_test(test_ace_from_str_invalid_access),
 		cmocka_unit_test(test_ace_from_str_multiple_flags),
 		cmocka_unit_test(test_ace_from_str_invalid_flags),
-		cmocka_unit_test(test_ace_from_str_perm_read_only),
-		cmocka_unit_test(test_ace_from_str_perm_none),
+		cmocka_unit_test(test_ace_from_str_perms),
 		cmocka_unit_test(test_ace_from_str_invalid_perms),
 		cmocka_unit_test(test_ace_from_str_empty_str),
 		cmocka_unit_test(test_ace_from_str_not_all_fields),
